orchestra: refactor code to remove redundancy
After I moved everything into the same package, there is lots of
redundancy that we can remove.

Part of https://github.com/ooni/probe-engine/issues/651.
