+++ internal/orchestra/register.go
@@ -2,49 +2,45 @@ package orchestra
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/ooni/probe-engine/internal/httpx"
-	"github.com/ooni/probe-engine/model"
 )
-
-// RegisterConfig contains configs for registering to OONI orchestra.
-type RegisterConfig struct {
-	BaseURL    string
-	HTTPClient *http.Client
-	Logger     model.Logger
-	Metadata   Metadata
-	Password   string
-	UserAgent  string
-}
 
 type registerRequest struct {
 	Metadata
 	Password string `json:"password"`
 }
 
-// RegisterResult contains the result of logging in.
-type RegisterResult struct {
+type registerResult struct {
 	ClientID string `json:"client_id"`
 }
 
-// Register registers this probe with OONI orchestra
-func Register(ctx context.Context, config RegisterConfig) (*RegisterResult, error) {
-	req := &registerRequest{
-		Metadata: config.Metadata,
-		Password: config.Password,
+// MaybeRegister registers this client if not already registered
+func (c Client) MaybeRegister(ctx context.Context, metadata Metadata) error {
+	if !metadata.Valid() {
+		return errInvalidMetadata
 	}
-	var resp RegisterResult
+	state := c.StateFile.Get()
+	if state.Credentials() != nil {
+		return nil // we're already good
+	}
+	c.RegisterCalls.Add(1)
+	pwd := randomPassword(64)
+	req := &registerRequest{
+		Metadata: metadata,
+		Password: pwd,
+	}
+	var resp registerResult
 	err := (httpx.Client{
-		BaseURL:    config.BaseURL,
-		HTTPClient: config.HTTPClient,
-		Logger:     config.Logger,
-		UserAgent:  config.UserAgent,
+		BaseURL:    c.BaseURL,
+		HTTPClient: c.HTTPClient,
+		Logger:     c.Logger,
+		UserAgent:  c.UserAgent,
 	}).CreateJSON(ctx, "/api/v1/register", req, &resp)
 	if err != nil {
-		return nil, err
+		return err
 	}
-	// Implementation note: the API does not return 200 unless there
-	// is success, so we don't bother with reading the error field
-	return &resp, nil
+	state.ClientID = resp.ClientID
+	state.Password = pwd
+	return c.StateFile.Set(state)
 }
