+++ internal/orchestra/register_test.go
@@ -2,37 +2,58 @@ package orchestra_test
 
 import (
 	"context"
-	"net/http"
 	"testing"
 
-	"github.com/apex/log"
+	"github.com/ooni/probe-engine/internal/mockable"
 	"github.com/ooni/probe-engine/internal/orchestra"
 )
 
-func TestRegisterSuccess(t *testing.T) {
-	clientID, err := Register()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if clientID == "" {
-		t.Fatal("ClientID should not be empty")
-	}
+func TestUnitMaybeRegister(t *testing.T) {
+	t.Run("when metadata is not valid", func(t *testing.T) {
+		clnt := newclient()
+		ctx := context.Background()
+		var metadata orchestra.Metadata
+		if err := clnt.MaybeRegister(ctx, metadata); err == nil {
+			t.Fatal("expected an error here")
+		}
+	})
+	t.Run("when we have already registered", func(t *testing.T) {
+		clnt := newclient()
+		state := orchestra.State{
+			ClientID: "xx-xxx-x-xxxx",
+			Password: "xx",
+		}
+		if err := clnt.StateFile.Set(state); err != nil {
+			t.Fatal(err)
+		}
+		ctx := context.Background()
+		metadata := mockable.OrchestraMetadataFixture()
+		if err := clnt.MaybeRegister(ctx, metadata); err != nil {
+			t.Fatal(err)
+		}
+	})
+	t.Run("when the API call fails", func(t *testing.T) {
+		clnt := newclient()
+		clnt.BaseURL = "\t\t\t"
+		ctx := context.Background()
+		metadata := mockable.OrchestraMetadataFixture()
+		if err := clnt.MaybeRegister(ctx, metadata); err == nil {
+			t.Fatal("expected an error here")
+		}
+	})
 }
 
-func TestRegisterFailure(t *testing.T) {
-	// The successful integration test contains the minimal amount
-	// of fields expected by the orchestra. Any less amount of fields,
-	// such as we do here, results in the API returning error.
-	result, err := orchestra.Register(context.Background(), orchestra.RegisterConfig{
-		BaseURL:    "https://ps-test.ooni.io",
-		HTTPClient: http.DefaultClient,
-		Logger:     log.Log,
-		UserAgent:  "miniooni/0.1.0-dev",
-	})
-	if err == nil {
-		t.Fatal("expected an error here")
+func TestIntegrationMaybeRegisterIdempotent(t *testing.T) {
+	clnt := newclient()
+	ctx := context.Background()
+	metadata := mockable.OrchestraMetadataFixture()
+	if err := clnt.MaybeRegister(ctx, metadata); err != nil {
+		t.Fatal(err)
 	}
-	if result != nil {
-		t.Fatal("result should be nil here")
+	if err := clnt.MaybeRegister(ctx, metadata); err != nil {
+		t.Fatal(err)
+	}
+	if clnt.RegisterCalls.Load() != 1 {
+		t.Fatal("called register API too many times")
 	}
 }
